Fix test failures due to differences between `Hash` and `HashWithoutLabels` with `stringlabels`
